feat: move custom cup button to a floting action button
